@@ -1,5 +1,6 @@
 [metadata]
 name = veloxchem
+version = attr: veloxchem.__version__
 description = VeloxChem: an electronic structure code
 keywords = electronic structure theory, response theory, computational spectroscopy
 license = LGPLv3+
@@ -16,20 +17,30 @@ classifiers =
     License :: OSI Approved :: GNU Lesser General Public License v3 or later (LGPLv3+)
     Programming Language :: Python
     Programming Language :: Python :: 3
-    Programming Language :: Python :: 3.6
     Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
+    Programming Language :: Python :: 3.9
+    Programming Language :: Python :: 3.10
     Programming Language :: C++
     Topic :: Scientific/Engineering :: Chemistry
-version = attr: veloxchem.__version__
 
 [options]
-python_requires = >= 3.6
+packages = veloxchem
+
+package_dir =
+   veloxchem=src/pymodule
+
+python_requires = >= 3.7
 
 setup_requires =
     mpi4py >= 3.0
     numpy >= 1.15.0
-    pybind11 >= 2.6
+    pybind11[global] >= 2.6
+    setuptools >=46.4.0
+    wheel
+    scikit-build >=0.13
+    cmake >= 3.18
+    ninja; platform_system!='Windows'
 
 install_requires =
     geometric @ git+https://github.com/leeping/geomeTRIC.git@e982ad6#egg=geometric
@@ -43,7 +54,7 @@ test_suite =
 
 [options.extras_require]
 qmmm =
-    cppe == 0.2.1
+    cppe >= 0.3.1
     MDAnalysis >= 1.0
 
 doc =
@@ -51,6 +62,9 @@ doc =
     sphinx_rtd_theme
 
 test =
+    h5py >= 2.8
+    geometric @ git+https://github.com/leeping/geomeTRIC.git@e982ad6#egg=geometric
+    psutil >= 5.6
     pytest
     pytest-cov
 
