@@ -1,20 +1,15 @@
 import random
-import sys
 from pathlib import Path
 
 import numpy as np
 import pytest
-
-try:
-    import cppe
-except ImportError:
-    pass
 
 from veloxchem.lreigensolver import LinearResponseEigenSolver
 from veloxchem.mpitask import MpiTask
 from veloxchem.scfrestdriver import ScfRestrictedDriver
 from veloxchem.veloxchemlib import hartree_in_ev, is_mpi_master
 
+from .addons import using_cppe
 
 @pytest.mark.solvers
 class TestRPA:
@@ -128,7 +123,7 @@ class TestRPA:
 
         self.run_rpa(inpfile, potfile, xcfun_label, data_lines)
 
-    @pytest.mark.skipif('cppe' not in sys.modules, reason='cppe not available')
+    @using_cppe
     def test_rpa_hf_pe(self):
 
         here = Path(__file__).parent
@@ -151,7 +146,7 @@ class TestRPA:
 
         self.run_rpa(inpfile, potfile, xcfun_label, data_lines)
 
-    @pytest.mark.skipif('cppe' not in sys.modules, reason='cppe not available')
+    @using_cppe
     def test_rpa_dft_pe(self):
 
         here = Path(__file__).parent
