@@ -43,13 +43,14 @@ class TestScfRestrictedOpenShell:
             ('heh.inp', None, 'def2-svp', -3.34847190869),
             ('heh.inp', 'slater', 'def2-svp', -3.166481549682),
             ('heh.inp', 'b3lyp', 'def2-svp', -3.404225946805),
+            ('water_triplet.inp', None, 'def2-svp', -75.701357737185),
             ('water_triplet.inp', 'blyp', 'def2-svp', -76.055256325587),
             ('water_triplet.inp', 'b3lyp', 'def2-svp', -76.074465451578),
             ('water_triplet.inp', 'b3lyp', 'aug-cc-pvdz', -76.176630915242),
         ],
         ids=[
             'Li', 'ClH2', 'H3O', 'CH5', 'HeH', 'HeH-Slater', 'HeH-B3LYP',
-            'H2O-BLYP', 'H2O-B3LYP', 'H2O-B3LYP-aDZ'
+            'H2O-HF', 'H2O-BLYP', 'H2O-B3LYP', 'H2O-B3LYP-aDZ'
         ],
         # TODO: add H2O-SLDA test
     )
