@@ -247,7 +247,8 @@ CAOKohnShamMatrix::collect(int32_t rank, int32_t nodes, MPI_Comm comm, int32_t s
 
             std::vector<int32_t> xc_integers(4);
 
-            auto merror = MPI_Recv(xc_integers.data(), xc_integers.size(), MPI_INT32_T, source, tag_id++, comm, &mstat);
+            auto merror = MPI_Recv(xc_integers.data(), static_cast<int32_t>(xc_integers.size()),
+                                   MPI_INT32_T, source, tag_id++, comm, &mstat);
 
             if (merror != MPI_SUCCESS) mpi::abort(merror, "collectKohnShamMatrix: xc_integers");
 
@@ -263,7 +264,8 @@ CAOKohnShamMatrix::collect(int32_t rank, int32_t nodes, MPI_Comm comm, int32_t s
 
             std::vector<double> xc_doubles(2);
 
-            merror = MPI_Recv(xc_doubles.data(), xc_doubles.size(), MPI_DOUBLE, source, tag_id++, comm, &mstat);
+            merror = MPI_Recv(xc_doubles.data(), static_cast<int32_t>(xc_doubles.size()),
+                              MPI_DOUBLE, source, tag_id++, comm, &mstat);
 
             if (merror != MPI_SUCCESS) mpi::abort(merror, "collectKohnShamMatrix: xc_doubles");
 
@@ -303,7 +305,8 @@ CAOKohnShamMatrix::collect(int32_t rank, int32_t nodes, MPI_Comm comm, int32_t s
 
             std::vector<int32_t> xc_integers({nxcmats, nrows, ncols, rest});
 
-            auto merror = MPI_Send(xc_integers.data(), xc_integers.size(), MPI_INT32_T, mpi::master(), tag_id++, comm);
+            auto merror = MPI_Send(xc_integers.data(), static_cast<int32_t>(xc_integers.size()),
+                                   MPI_INT32_T, mpi::master(), tag_id++, comm);
 
             if (merror != MPI_SUCCESS) mpi::abort(merror, "collectKohnShamMatrix: xc_integers");
 
@@ -311,7 +314,8 @@ CAOKohnShamMatrix::collect(int32_t rank, int32_t nodes, MPI_Comm comm, int32_t s
 
             std::vector<double> xc_doubles({_xcElectrons, _xcEnergy});
 
-            merror = MPI_Send(xc_doubles.data(), xc_doubles.size(), MPI_DOUBLE, mpi::master(), tag_id++, comm);
+            merror = MPI_Send(xc_doubles.data(), static_cast<int32_t>(xc_doubles.size()),
+                              MPI_DOUBLE, mpi::master(), tag_id++, comm);
 
             if (merror != MPI_SUCCESS) mpi::abort(merror, "collectKohnShamMatrix: xc_doubles");
 
