@@ -33,12 +33,9 @@ CXCGradientGrid::CXCGradientGrid()
 
     , _xcValues(CMemBlock2D<double>())
 {
-    
 }
 
-CXCGradientGrid::CXCGradientGrid(const CMemBlock2D<double>& xcValues,
-                                 const dengrid              densityGridType,
-                                 const xcfun                xcGridType)
+CXCGradientGrid::CXCGradientGrid(const CMemBlock2D<double>& xcValues, const dengrid densityGridType, const xcfun xcGridType)
 
     : _densityGridType(densityGridType)
 
@@ -46,27 +43,23 @@ CXCGradientGrid::CXCGradientGrid(const CMemBlock2D<double>& xcValues,
 
     , _xcValues(xcValues)
 {
-    
 }
 
-
-CXCGradientGrid::CXCGradientGrid(const int32_t nGridPoints,
-                                 const dengrid densityGridType,
-                                 const xcfun   xcGridType)
+CXCGradientGrid::CXCGradientGrid(const int32_t nGridPoints, const dengrid densityGridType, const xcfun xcGridType)
 {
     _densityGridType = densityGridType;
-    
+
     _xcGridType = xcGridType;
-    
+
     int32_t ncomp = 0;
-    
+
     if (_xcGridType == xcfun::lda) ncomp = 3;
-    
+
     if (_xcGridType == xcfun::gga) ncomp = 6;
-    
+
     // NOTE: this needs to be checked with mgga functionals implementation
     if (_xcGridType == xcfun::mgga) ncomp = (_densityGridType == dengrid::ab) ? 8 : 4;
-    
+
     _xcValues = CMemBlock2D<double>(nGridPoints, ncomp);
 }
 
@@ -98,13 +91,13 @@ CXCGradientGrid&
 CXCGradientGrid::operator=(const CXCGradientGrid& source)
 {
     if (this == &source) return *this;
-    
+
     _densityGridType = source._densityGridType;
-    
+
     _xcGridType = source._xcGridType;
-    
+
     _xcValues = source._xcValues;
-    
+
     return *this;
 }
 
@@ -112,13 +105,13 @@ CXCGradientGrid&
 CXCGradientGrid::operator=(CXCGradientGrid&& source) noexcept
 {
     if (this == &source) return *this;
-    
+
     _densityGridType = std::move(source._densityGridType);
-    
+
     _xcGridType = std::move(source._xcGridType);
-    
+
     _xcValues = std::move(source._xcValues);
-    
+
     return *this;
 }
 
@@ -126,11 +119,11 @@ bool
 CXCGradientGrid::operator==(const CXCGradientGrid& other) const
 {
     if (_densityGridType != other._densityGridType) return false;
-    
+
     if (_xcGridType != other._xcGridType) return false;
-    
+
     if (_xcValues != other._xcValues) return false;
-    
+
     return true;
 }
 
@@ -171,28 +164,36 @@ CXCGradientGrid::xcGradientValues(const xcvars gradientType) const
     {
         return _xcValues.data(1);
     }
-    
+
     if (gradientType == xcvars::rhob)
     {
         return _xcValues.data(2);
     }
-    
+
     if (gradientType == xcvars::grada)
     {
         return _xcValues.data(3);
     }
-    
+
     if (gradientType == xcvars::gradb)
     {
         return _xcValues.data(4);
-
     }
-    
+
     if (gradientType == xcvars::gradab)
     {
         return _xcValues.data(5);
     }
-    
+
+    if (gradientType == xcvars::taua)
+    {
+        return _xcValues.data(6);
+    }
+
+    if (gradientType == xcvars::taub)
+    {
+        return _xcValues.data(7);
+    }
     return nullptr;
 }
 
@@ -203,28 +204,37 @@ CXCGradientGrid::xcGradientValues(const xcvars gradientType)
     {
         return _xcValues.data(1);
     }
-    
+
     if (gradientType == xcvars::rhob)
     {
-         return _xcValues.data(2);
+        return _xcValues.data(2);
     }
-    
+
     if (gradientType == xcvars::grada)
     {
-         return _xcValues.data(3);
+        return _xcValues.data(3);
     }
-    
+
     if (gradientType == xcvars::gradb)
     {
-       return _xcValues.data(4);
+        return _xcValues.data(4);
     }
-    
+
     if (gradientType == xcvars::gradab)
     {
         return _xcValues.data(5);
-
     }
-    
+
+    if (gradientType == xcvars::taua)
+    {
+        return _xcValues.data(6);
+    }
+
+    if (gradientType == xcvars::taub)
+    {
+        return _xcValues.data(7);
+    }
+
     return nullptr;
 }
 
@@ -238,16 +248,16 @@ std::ostream&
 operator<<(std::ostream& output, const CXCGradientGrid& source)
 {
     output << std::endl;
-    
+
     output << "[CXCGradientGrid (Object):" << &source << "]" << std::endl;
-    
+
     output << "_densityGridType: " << to_string(source._densityGridType) << std::endl;
-    
+
     output << "_xcGridType: " << to_string(source._xcGridType) << std::endl;
-    
+
     output << "_xcValues: " << std::endl;
-    
+
     output << source._xcValues << std::endl;
-    
+
     return output;
 }
