@@ -33,6 +33,7 @@
 #include "Codata.hpp"
 #include "MathFunc.hpp"
 #include "StringFormat.hpp"
+#include "ChemicalElement.hpp"
 
 CMolecule::CMolecule(const std::vector<double>&      atomCoordinates,
                      const std::vector<double>&      atomCharges,
@@ -66,6 +67,8 @@ CMolecule::CMolecule(const CMolecule& source)
 
     , _atomLabels(source._atomLabels)
 
+    , _idsAtomic(source._idsAtomic)
+
     , _idsElemental(source._idsElemental)
 {
 }
@@ -83,6 +86,8 @@ CMolecule::CMolecule(CMolecule&& source) noexcept
     , _atomMasses(std::move(source._atomMasses))
 
     , _atomLabels(std::move(source._atomLabels))
+
+    , _idsAtomic(std::move(source._idsAtomic))
 
     , _idsElemental(std::move(source._idsElemental))
 {
@@ -172,12 +177,17 @@ CMolecule::CMolecule(const CMolecule& mol_1, const CMolecule& mol_2)
 }
 
 CMolecule
-CMolecule::getSubMolecule(int32_t startIndex, int32_t numAtoms)
+CMolecule::getSubMolecule(int32_t startIndex,
+                          int32_t numAtoms)
 {
     std::vector<double>      atomCoordinates;
+    
     std::vector<double>      atomCharges;
+    
     std::vector<double>      atomMasses;
+    
     std::vector<std::string> atomLabels;
+    
     std::vector<int32_t>     idsElemental;
 
     // boundary check
@@ -241,6 +251,8 @@ CMolecule::operator=(const CMolecule& source)
 
     _atomLabels = source._atomLabels;
 
+    _idsAtomic = source._idsAtomic;
+
     _idsElemental = source._idsElemental;
 
     return *this;
@@ -262,6 +274,8 @@ CMolecule::operator=(CMolecule&& source) noexcept
     _atomMasses = std::move(source._atomMasses);
 
     _atomLabels = std::move(source._atomLabels);
+
+    _idsAtomic = std::move(source._idsAtomic);
 
     _idsElemental = std::move(source._idsElemental);
 
@@ -288,6 +302,8 @@ CMolecule::operator==(const CMolecule& other) const
         if (_atomLabels[i] != other._atomLabels[i]) return false;
     }
 
+    if (_idsAtomic != other._idsAtomic) return false;
+
     if (_idsElemental != other._idsElemental) return false;
 
     return true;
@@ -297,6 +313,104 @@ bool
 CMolecule::operator!=(const CMolecule& other) const
 {
     return !(*this == other);
+}
+
+void
+CMolecule::addAtom(const std::string& atomLabel,
+                   const double       atomCoordinateX,
+                   const double       atomCoordinateY,
+                   const double       atomCoordinateZ)
+{
+    CChemicalElement elem;
+    
+    if (elem.setAtomType(atomLabel))
+    {
+        std::vector<double> atomCoordinates;
+        
+        std::vector<double> atomCharges;
+        
+        std::vector<double> atomMasses;
+        
+        std::vector<std::string> atomLabels;
+        
+        std::vector<int32_t> idsElemental;
+        
+        // set up pointers to atomic coordinates anf charges
+        
+        const auto natoms = getNumberOfAtoms();
+
+        auto coordx = getCoordinatesX();
+        
+        for (int32_t i = 0; i < natoms; i++)
+        {
+            atomCoordinates.push_back(coordx[i]);
+        }
+        
+        atomCoordinates.push_back(atomCoordinateX);
+        
+        auto coordy = getCoordinatesY();
+        
+        for (int32_t i = 0; i < natoms; i++)
+        {
+            atomCoordinates.push_back(coordy[i]);
+        }
+        
+        atomCoordinates.push_back(atomCoordinateY);
+        
+        auto coordz = getCoordinatesZ();
+        
+        for (int32_t i = 0; i < natoms; i++)
+        {
+            atomCoordinates.push_back(coordz[i]);
+        }
+        
+        atomCoordinates.push_back(atomCoordinateZ);
+        
+        for (int32_t i = 0; i < natoms; i++)
+        {
+            atomCharges.push_back(_atomCharges.at(i));
+
+            atomMasses.push_back(_atomMasses.at(i));
+
+            atomLabels.push_back(_atomLabels[i]);
+
+            idsElemental.push_back(_idsElemental.at(i));
+        }
+        
+        atomCharges.push_back(elem.getAtomicCharge());
+        
+        atomMasses.push_back(elem.getAtomicMass());
+        
+        atomLabels.push_back(atomLabel);
+        
+        idsElemental.push_back(elem.getIdentifier());
+        
+        *this = CMolecule(atomCoordinates, atomCharges, atomMasses,
+                          atomLabels, idsElemental);
+    }
+}
+
+void
+CMolecule::setAtomicIndexes(const int32_t startIndex)
+{
+    for (int32_t i = 0; i < _idsAtomic.size(); i++)
+    {
+        _idsAtomic.at(i) = startIndex + i;
+    }
+}
+
+std::vector<double>
+CMolecule::getAtomCoordinates(const int32_t iAtom) const
+{
+    std::vector<double> coords;
+    
+    coords.push_back(getCoordinatesX()[iAtom]);
+    
+    coords.push_back(getCoordinatesY()[iAtom]);
+    
+    coords.push_back(getCoordinatesZ()[iAtom]);
+    
+    return coords;
 }
 
 void
@@ -557,6 +671,170 @@ CMolecule::getLabel(const int32_t iAtom) const
     return std::string();
 }
 
+std::vector<int32_t>
+CMolecule::getAtomIndexes(const std::string& atomLabel) const
+{
+    CChemicalElement elem;
+    
+    if (elem.setAtomType(atomLabel))
+    {
+        std::vector<int32_t> indexes;
+        
+        const auto idatm = elem.getIdentifier();
+        
+        for (int32_t i = 0; i < getNumberOfAtoms(); i++)
+        {
+            if (_idsElemental.at(i) == idatm)
+            {
+                indexes.push_back(i);
+            }
+        }
+        
+        return indexes;
+    }
+    
+    return std::vector<int32_t>();
+}
+
+int32_t
+CMolecule::getIndexOfNearestAtom(const int32_t iAtom) const
+{
+    // set up pointers to coordinates
+    
+    auto coordx = getCoordinatesX();
+    
+    auto coordy = getCoordinatesY();
+    
+    auto coordz = getCoordinatesZ();
+    
+    // set up coordinates of reference atom
+    
+    const auto rax = coordx[iAtom];
+    
+    const auto ray = coordy[iAtom];
+    
+    const auto raz = coordz[iAtom];
+    
+    // determine position of nearest atom
+    
+    int32_t idx = -1;
+    
+    double rmin = 1.0e9;
+    
+    for (int32_t i = 0; i < getNumberOfAtoms(); i++)
+    {
+        if (i != iAtom)
+        {
+            const double rab = mathfunc::distance(rax, ray, raz,
+                                                  coordx[i],
+                                                  coordy[i],
+                                                  coordz[i]);
+            
+            if (rab < rmin)
+            {
+                rmin = rab;
+                
+                idx = i;
+            }
+        }
+    }
+    
+    return idx;
+}
+
+int32_t
+CMolecule::getCoordinationNummber(const int32_t iAtom,
+                                  const double  radius) const
+{
+    const auto natoms = getNumberOfAtoms();
+    
+    if (natoms > 0)
+    {
+        // set up pointers to coordinates
+           
+        auto coordx = getCoordinatesX();
+           
+        auto coordy = getCoordinatesY();
+           
+        auto coordz = getCoordinatesZ();
+        
+        // set up atom coordinates
+        
+        const auto rax = coordx[iAtom];
+        
+        const auto ray = coordy[iAtom];
+        
+        const auto raz = coordz[iAtom];
+        
+        int32_t cnum = 0;
+        
+        for (int32_t i = 0; i < natoms; i++)
+        {
+            if (i != iAtom)
+            {
+                const auto rab = mathfunc::distance(rax, ray, raz,
+                                                    coordx[i],
+                                                    coordy[i],
+                                                    coordz[i]);
+                
+                if (rab < radius)
+                {
+                    cnum++;
+                }
+            }
+        }
+        
+        return cnum;
+    }
+    
+    return 0;
+}
+
+double
+CMolecule::getMinDistance(const double coordinateX,
+                          const double coordinateY,
+                          const double coordinateZ) const
+{
+    const auto natoms = getNumberOfAtoms();
+    
+    if (natoms > 0)
+    {
+        // set up pointers to coordinates
+           
+        auto coordx = getCoordinatesX();
+           
+        auto coordy = getCoordinatesY();
+           
+        auto coordz = getCoordinatesZ();
+        
+        double rmin = mathfunc::distance(coordinateX,
+                                         coordinateY,
+                                         coordinateZ,
+                                         coordx[0],
+                                         coordy[0],
+                                         coordz[0]);
+        
+        for (int32_t i = 1; i < natoms; i++)
+        {
+            const auto rab = mathfunc::distance(coordinateX,
+                                                coordinateY,
+                                                coordinateZ,
+                                                coordx[i],
+                                                coordy[i],
+                                                coordz[i]);
+            
+            if (rmin > rab)
+            {
+                rmin = rab;
+            }
+        }
+        
+        return rmin;
+    }
+    
+    return -1.0;
+}
+
 std::string
 CMolecule::printGeometry() const
 {
@@ -654,6 +932,8 @@ CMolecule::broadcast(int32_t rank, MPI_Comm comm)
 
         mpi::bcast(_atomLabels, rank, comm);
 
+        _idsAtomic.broadcast(rank, comm);
+
         _idsElemental.broadcast(rank, comm);
     }
 }
@@ -683,6 +963,8 @@ operator<<(std::ostream& output, const CMolecule& source)
 
         output << source._atomLabels[i] << std::endl;
     }
+
+    output << "_idsAtomic: " << source._idsAtomic << std::endl;
 
     output << "_idsElemental: " << source._idsElemental << std::endl;
 
