@@ -77,6 +77,11 @@ class CMolecule
     std::vector<std::string> _atomLabels;
 
     /**
+     The vector of global indexes.
+     */
+    CMemBlock<int32_t> _idsAtomic;
+
+    /**
      The vector of chemical element identifiers.
      */
     CMemBlock<int32_t> _idsElemental;
@@ -122,7 +127,8 @@ class CMolecule
      @param mol_1 the first molecule object.
      @param mol_2 the second molecule object.
      */
-    CMolecule(const CMolecule& mol_1, const CMolecule& mol_2);
+    CMolecule(const CMolecule& mol_1,
+              const CMolecule& mol_2);
 
     /**
      Creates a sub-molecule object by slicing the molecule object.
@@ -130,7 +136,22 @@ class CMolecule
      @param startIndex the starting index of the sub-molecule (0-based).
      @param numAtoms the number of atoms in the sub-molecule.
      */
-    CMolecule getSubMolecule(int32_t startIndex, int32_t numAtoms);
+    CMolecule getSubMolecule(int32_t startIndex,
+                             int32_t numAtoms);
+    
+    
+    /**
+     Adds atom to molecule using given atom label and coordinates.
+
+     @param atomLabel the label of atom.
+     @param atomCoordinateX the coordinate X of atom.
+     @param atomCoordinateY the coordinate Y of atom.
+     @param atomCoordinateZ the coordinate Z of atom.
+    */
+    void addAtom(const std::string& atomLabel,
+                 const double       atomCoordinateX,
+                 const double       atomCoordinateY,
+                 const double       atomCoordinateZ);
 
     /**
      Assigns a molecule object by copying other molecule object.
@@ -161,6 +182,13 @@ class CMolecule
      @return true if molecule objects are not equal, false otherwise.
      */
     bool operator!=(const CMolecule& other) const;
+
+    /**
+     Sets atomic indexes of all atoms in molecule object.
+
+     @param startIndex the starting index of atomic indexes.
+     */
+    void setAtomicIndexes(const int32_t startIndex);
 
     /**
      Sets charge of molecule object.
@@ -214,7 +242,9 @@ class CMolecule
     @param idElemental the chemical element number.
     @return the number of atoms.
     */
-    int32_t getNumberOfAtoms(const int32_t iAtom, const int32_t nAtoms, const int32_t idElemental) const;
+    int32_t getNumberOfAtoms(const int32_t iAtom,
+                             const int32_t nAtoms,
+                             const int32_t idElemental) const;
 
     /**
      Gets set of unique chemical elements in molecule.
@@ -285,6 +315,19 @@ class CMolecule
      @return the vector of distances.
      */
     CMemBlock<double> getMinDistances() const;
+    
+    /**
+     Computes minimal distance from given external point (x,y,z) to closest atom
+     in molecule.
+
+     @param coordinateX the coordinate X of external point.
+     @param coordinateY the coordinate Y of external point.
+     @param coordinateZ the coordinate Z of external point.
+     @return the minimal distance.
+    */
+    double getMinDistance(const double coordinateX,
+                          const double coordinateY,
+                          const double coordinateZ) const; 
 
     /**
      Gets nuclear repulsion energy for molecule assuming point charge model for
@@ -316,12 +359,47 @@ class CMolecule
     std::vector<double> getCovalentRadii() const;
 
     /**
-     Gets labele of specific atom.
+     Gets label of specific atom.
 
      @param iAtom the index of atom.
      @return the label of atom.
      */
     std::string getLabel(const int32_t iAtom) const;
+    
+    /**
+     Gets indexes of atoms with given atomic label.
+
+     @param atomLabel the label of requested atom type.
+     @return the vector of atom indexes.
+    */
+    std::vector<int32_t> getAtomIndexes(const std::string& atomLabel) const;
+    
+    /**
+     Gets vector of atom coordinates of specific atom.
+
+     @param iAtom the index of atom.
+     @return the vector of atom indexes.
+    */
+    std::vector<double> getAtomCoordinates(const int32_t iAtom) const;
+    
+    /**
+     Gets index of nearest atom with given atom label to specific atom.
+
+     @param iAtom the index of requested atom.
+     @return the index of nearest atom to requested atom.
+    */
+    int32_t getIndexOfNearestAtom(const int32_t iAtom) const;
+    
+    
+    /**
+     Gets coordination number of specific atom in molecule.
+
+     @param iAtom the index of requested atom.
+     @param raddius the effective coordination radius.
+     @return the coordination number of atom.
+    */
+    int32_t getCoordinationNummber(const int32_t iAtom,
+                                   const double  radius) const;
 
     /**
      Prints geometry of molecule as table to output stream.
@@ -355,7 +433,8 @@ class CMolecule
      @param output the output text stream.
      @param source the molecule object.
      */
-    friend std::ostream& operator<<(std::ostream& output, const CMolecule& source);
+    friend std::ostream& operator<<(      std::ostream& output,
+                                    const CMolecule&    source);
 };
 
 #endif /* Molecule_hpp */
