@@ -248,7 +248,8 @@ class CSADGuessDriver
      @param basis the molecular basis set.
      @return a vector of vector containing the atomic orbital indicies for each atom in the molecule.
      */
-    std::vector<std::vector<int32_t>> getAOIndicesOfAtoms(const CMolecule& molecule, const CMolecularBasis& basis) const;
+    std::vector<std::vector<int32_t>> getAOIndicesOfAtoms(const CMolecule&       molecule,
+                                                          const CMolecularBasis& basis) const;
 
     /**
      Computes occupation numbers for a given element.
@@ -257,7 +258,8 @@ class CSADGuessDriver
      @param nelec the number of excessive alpha or beta electrons.
      @return a vector of occupation numbers.
      */
-    std::vector<double> getOccupationNumbersForElement(const int32_t elem_id, const double nelec) const;
+    std::vector<double> getOccupationNumbersForElement(const int32_t elem_id,
+                                                       const double  nelec) const;
 
     /**
      Computes occupation numbers for a given molecule.
@@ -266,7 +268,8 @@ class CSADGuessDriver
      @param nelec the number of excessive alpha or beta electrons.
      @return a vector of vector containing the occupation numbers for each atom in the molecule.
      */
-    std::vector<std::vector<double>> getOccupationNumbersForMolecule(const CMolecule& molecule, const double nelec) const;
+    std::vector<std::vector<double>> getOccupationNumbersForMolecule(const CMolecule& molecule,
+                                                                     const double     nelec) const;
 };
 
 #endif /* SADGuessDriver_hpp */
