@@ -41,6 +41,7 @@ from .veloxchemlib import DenseMatrix
 from .veloxchemlib import TwoIndexes
 from .veloxchemlib import MOIntsBatch
 from .veloxchemlib import ExcitationVector
+from .veloxchemlib import XCIntegrator
 
 # C++ functions
 from .veloxchemlib import available_functionals
@@ -58,6 +59,7 @@ from .veloxchemlib import rotatory_strength_in_cgs
 from .veloxchemlib import extinction_coefficient_from_beta
 from .veloxchemlib import fine_structure_constant
 from .veloxchemlib import mathconst_pi
+from .veloxchemlib import parse_xc_func
 
 # C++ enums
 from .veloxchemlib import denmat
@@ -128,10 +130,12 @@ from .errorhandler import assert_msg_critical
 from .qqscheme import get_qq_type
 from .qqscheme import get_qq_scheme
 
-# Environment variable: basis set path and number of OpenMP threads
-from .environ import set_vlxbasispath, set_omp_num_threads, get_basis_path
+# Environment variable: basis set path, number of OpenMP threads, MKL linking
+from .environ import set_vlxbasispath, set_omp_num_threads, get_basis_path, configure_mkl_rt
 
 set_vlxbasispath()
 set_omp_num_threads()
+configure_mkl_rt()
+
 
 __version__ = "1.0rc2"
