@@ -38,6 +38,8 @@ from .respchargesdriver import RespChargesDriver
 from .excitondriver import ExcitonModelDriver
 from .numerovdriver import NumerovDriver
 from .mp2driver import Mp2Driver
+from .cnadriver import CnaAnalysisDriver
+from .gopdriver import GlobalOptimizationDriver
 from .loprop import LoPropDriver
 from .trajectorydriver import TrajectoryDriver
 from .scfgradientdriver import ScfGradientDriver
@@ -60,6 +62,7 @@ from .veloxchemlib import DiagEriDriver
 from .cli import cli
 from .errorhandler import assert_msg_critical
 from .slurminfo import get_slurm_end_time
+
 
 def select_scf_driver(task, scf_type):
     """
@@ -498,15 +501,39 @@ def main():
             chg_drv.compute(task.molecule, task.ao_basis, 'resp')
         elif task_type == 'esp charges':
             chg_drv.compute(task.molecule, task.ao_basis, 'esp')
-            
+
     # Test of electron repulstion integrals
-    
+
     if task_type == 'eritest':
         print('*** Testing Two Electron Implementation ***')
         tm0 = tm.time()
         eri_driver = DiagEriDriver()
         pgblock = eri_driver.compute(task.molecule, task.ao_basis)
         print('Diagonal Eri Driver: ', tm.time() - tm0, ' sec.')
+
+    # CNA correlation analysis
+
+    if task_type == 'cna':
+        if 'cna' in task.input_dict:
+            cna_dict = task.input_dict['cna']
+        else:
+            cna_dict = {}
+
+        cna_drv = CnaAnalysisDriver(task.mpi_comm, task.ostream)
+        cna_drv.update_settings(cna_dict)
+        cna_drv.compute()
+
+    # Global optimization with tree-growth scheme
+
+    if task_type == 'gop':
+        if 'gop' in task.input_dict:
+            gop_dict = task.input_dict['gop']
+        else:
+            gop_dict = {}
+
+        gop_drv = GlobalOptimizationDriver(task.mpi_comm, task.ostream)
+        gop_drv.update_settings(gop_dict)
+        gop_drv.compute(task.input_dict['filename'], task.molecule)
 
     # All done
 
