@@ -23,22 +23,23 @@
 #  You should have received a copy of the GNU Lesser General Public License
 #  along with VeloxChem. If not, see <https://www.gnu.org/licenses/>.
 
-from mpi4py import MPI
-import numpy as np
 from pathlib import Path
+
+import numpy as np
+from mpi4py import MPI
+
 try:
     import cppe
 except ImportError:
     raise ImportError('Unable to import cppe. Please install cppe via ' +
-                      '\'python3 -m pip install cppe==0.2.1\'')
+                      '\'python3 -m pip install cppe\'')
 
-from .veloxchemlib import NuclearPotentialIntegralsDriver
-from .veloxchemlib import ElectricFieldIntegralsDriver
-from .veloxchemlib import bohr_in_angstroms
-from .veloxchemlib import mpi_master
+from .errorhandler import assert_msg_critical
 from .inputparser import InputParser
 from .subcommunicators import SubCommunicators
-from .errorhandler import assert_msg_critical
+from .veloxchemlib import (ElectricFieldIntegralsDriver,
+                           NuclearPotentialIntegralsDriver, bohr_in_angstroms,
+                           mpi_master)
 
 
 class PolEmbed:
@@ -82,7 +83,6 @@ class PolEmbed:
         self.output = ''
         self.V_es = None
 
-        self.check_cppe_version()
         self.update_options(pe_dict)
 
         cppe_mol = cppe.Molecule()
@@ -105,19 +105,6 @@ class PolEmbed:
                 if site.is_polarizable
             ])
             self.polarizable_coords = coords
-
-    def check_cppe_version(self):
-        """
-        Checks the version of CPPE.
-        """
-
-        from pkg_resources import parse_version
-        cppe_version = '0.2.1'
-
-        if parse_version(cppe.__version__) != parse_version(cppe_version):
-            err_str = 'cppe version {} is required. '.format(cppe_version)
-            err_str += 'cppe version {} was found.'.format(cppe.__version__)
-            raise ModuleNotFoundError(err_str)
 
     def get_cppe_version(self):
         """
