@@ -156,8 +156,8 @@ class QuadraticResponseDriver(NonLinearSolver):
 
         if self.rank == mpi_master():
             S = scf_tensors['S']
-            da = scf_tensors['D'][0]
-            mo = scf_tensors['C']
+            da = scf_tensors['D_alpha']
+            mo = scf_tensors['C_alpha']
             d_a_mo = np.linalg.multi_dot([mo.T, S, da, S, mo])
             norb = mo.shape[1]
         else:
@@ -296,14 +296,10 @@ class QuadraticResponseDriver(NonLinearSolver):
         """
 
         if self.rank == mpi_master():
-            S = scf_tensors['S']
-            D0 = scf_tensors['D'][0]
-            mo = scf_tensors['C']
-            F0 = np.linalg.multi_dot([mo.T, scf_tensors['F'][0], mo])
+            mo = scf_tensors['C_alpha']
+            F0 = np.linalg.multi_dot([mo.T, scf_tensors['F_alpha'], mo])
             norb = mo.shape[1]
         else:
-            S = None
-            D0 = None
             mo = None
             F0 = None
             norb = None
@@ -317,7 +313,7 @@ class QuadraticResponseDriver(NonLinearSolver):
         # computing all compounded first-order densities
         if self.rank == mpi_master():
             first_order_dens, second_order_dens = self.get_densities(
-                freqpairs, kX, S, D0, mo)
+                freqpairs, kX, mo, nocc)
         else:
             first_order_dens = None
             second_order_dens = None
@@ -393,23 +389,24 @@ class QuadraticResponseDriver(NonLinearSolver):
 
         return result
 
-    def get_densities(self, freqpairs, kX, S, D0, mo):
+    def get_densities(self, freqpairs, kX, mo, nocc):
         """
-        Computes the  densities needed for the perturbed Fock matrices.
+        Computes the densities needed for the perturbed Fock matrices.
 
-        :param wi:
+        :param freqpairs:
             A list of the frequencies
         :param kX:
             A dictonary with all the first-order response matrices
-        :param S:
-            The overlap matrix
-        :param D0:
-            The SCF density matrix in AO basis
         :param mo:
             A matrix containing the MO coefficents
+        :param nocc:
+            Number of occupied orbitals
 
         :return:
-            A list of tranformed densities
+            first_order_dens:
+             A list of first-order one-time tranformed compounded densities
+            second_order_dens:
+             A list of first-order two-time tranformed compounded densities
         """
 
         first_order_dens = []
@@ -417,27 +414,30 @@ class QuadraticResponseDriver(NonLinearSolver):
 
         for (wb, wc) in freqpairs:
 
-            # convert response matrix to ao basis #
-
-            kb = self.mo2ao(mo, kX[('B', wb)])
-            kc = self.mo2ao(mo, kX[('C', wc)])
+            kb = kX[('B', wb)]
+            kc = kX[('C', wc)]
 
             # create the first order single indexed densiteies #
 
-            Db = self.transform_dens(kb, D0, S)
-            Dc = self.transform_dens(kc, D0, S)
+            Db = self.commut_mo_density(kb, nocc)
+            Dc = self.commut_mo_density(kc, nocc)
 
             # create the first order two indexed densities #
 
-            Dbc = self.transform_dens(kb, Dc, S)
-            Dcb = self.transform_dens(kc, Db, S)
+            Dbc = self.commut(kb, Dc) + self.commut(kc, Db)
+
+            # density transformation from MO to AO basis
+
+            Db = np.linalg.multi_dot([mo, Db, mo.T])
+            Dc = np.linalg.multi_dot([mo, Dc, mo.T])
+            Dbc = np.linalg.multi_dot([mo, Dbc, mo.T])
 
             first_order_dens.append(Db.real)
             first_order_dens.append(Db.imag)
             first_order_dens.append(Dc.real)
             first_order_dens.append(Dc.imag)
-            second_order_dens.append((Dbc + Dcb).real)
-            second_order_dens.append((Dbc + Dcb).imag)
+            second_order_dens.append(Dbc.real)
+            second_order_dens.append(Dbc.imag)
 
         return first_order_dens, second_order_dens
 
